apply title ticker to channel names
